% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescent_R.R
\name{betahat_GD_R}
\alias{betahat_GD_R}
\title{Estimation of Linear Regression model via Gradient Descent method}
\usage{
betahat_GD_R(beta, x, y, tolerance, maxit, stepsize)
}
\arguments{
\item{beta}{[numeric] vector containing an initial guess for the beta vector}

\item{x}{[numeric] design matrix}

\item{y}{[numeric] response variable vector}

\item{tolerance}{[numeric] tolerance level, stopping criteria of the algorithm (error<tolerance: stop)}

\item{maxit}{[numeric] maximum number of iterations, used if the stopping criteria is never matched}

\item{stepsize}{[numeric] learning parameter, to update the parameters at each iteration}
}
\value{
[numeric] vector of beta parameters estimated
}
\description{
betahat_GD_R is used to estimates the beta vector of a linear Regression model via 
an unconstrained optimization method called Gradient Descent Method
}
