% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betahat_SD_Cpp}
\alias{betahat_SD_Cpp}
\title{betahat_SD_Cpp}
\usage{
betahat_SD_Cpp(beta, X, Y, tolerance, maxit)
}
\arguments{
\item{beta}{[numeric] vector containing an initial guess for the beta vector}

\item{X}{[numeric] design matrix}

\item{Y}{[numeric] response variable vector}

\item{tolerance}{[numeric] tolerance level, stopping criteria of the algorithm 
(error<tolerance: stop)}

\item{maxit}{[numeric] maximum number of iterations, used if the stopping 
criteria is never matched}
}
\description{
Estimation Linear Regression model via Steepest Descend method
Output: [list] Beta paraameters estimated
}
