% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteepestDescent_R.R
\name{betahat_SD_R}
\alias{betahat_SD_R}
\title{Steepest Descent Optimization Method for Linear Models}
\usage{
betahat_SD_R(beta, x, y, tolerance, maxit)
}
\arguments{
\item{beta}{[numeric] vector containing an initial guess for the beta vector}

\item{x}{[numeric] design matrix}

\item{y}{[numeric] response variable vector}

\item{tolerance}{[numeric] tolerance level, stopping criteria of the algorithm (error<tolerance: stop)}

\item{maxit}{[numeric] maximum number of iterations, used if the stopping criteria is never matched}
}
\value{
[numeric] vector of beta parameters estimated
}
\description{
betahat_SD_R is used to estimates the beta vector of a linear Regression model via Steepest Descent
}
