% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD_k fold cv_parallel.R
\name{cvparallel}
\alias{cvparallel}
\title{Title
## using parallel method to calculate the k-fold crossing validation}
\usage{
cvparallel(k, formula, df, tolerance, maxit, stepsize, verbose, cores)
}
\arguments{
\item{k}{k-fold cross validation}

\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}

\item{stepsize}{learning parameter}

\item{verbose}{should the function write messages during the computation?}

\item{cores}{the number of cores to be used in the parallelization}
}
\value{

}
\description{
Title
## using parallel method to calculate the k-fold crossing validation
}
