% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD_k fold cv_sequential.R
\name{cvsequencial}
\alias{cvsequencial}
\title{Title cvsequencial}
\usage{
cvsequencial(k, formula, df, tolerance, maxit, stepsize, verbose)
}
\arguments{
\item{k}{k-fold cross validation}

\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}

\item{stepsize}{learning parameter}

\item{verbose}{should the function write messages during the computation?}
}
\value{

}
\description{
## using sequential method to calculate the k-fold crossing validation
}
