% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_xy.R
\name{data_xy}
\alias{data_xy}
\title{Dataset in the form needed for estimating Linear Models}
\usage{
data_xy(formula, df)
}
\arguments{
\item{formula}{[character] a formula specifying the model}

\item{df}{[numeric] dataset containing the variables of interest}
}
\value{
[list] matrix of regressors and a vector of the response variable
}
\description{
Dataset in the form needed for estimating Linear Models
}
