% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_specification.R
\name{lm.model_spec}
\alias{lm.model_spec}
\title{Title}
\usage{
lm.model_spec(formula, df)
}
\arguments{
\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}
}
\value{

}
\description{
Title
}
