% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD_k fold cv_parallel.R, R/GD_k fold
%   cv_sequential.R, R/lm_GD_optimizer.R
\name{lm_GD_optimizer}
\alias{lm_GD_optimizer}
\title{Title
## using parallel method to calculate the k-fold crossing validation}
\usage{
lm_GD_optimizer(formula, df, tolerance, maxit, stepsize, verbose)

lm_GD_optimizer(formula, df, tolerance, maxit, stepsize, verbose)

lm_GD_optimizer(formula, df, tolerance, maxit, stepsize, verbose)
}
\arguments{
\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}

\item{stepsize}{learning parameter}

\item{verbose}{should the function write messages during the computation?}

\item{k}{k-fold cross validation}
}
\value{





}
\description{
## using sequential method to calculate the k-fold crossing validation
}
