% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_SD_optimizer.R
\name{lm_SD_optimizer}
\alias{lm_SD_optimizer}
\title{Title}
\usage{
lm_SD_optimizer(formula, df, tolerance, maxit, verbose)
}
\arguments{
\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}

\item{verbose}{should the function write messages during the computation?}
}
\value{

}
\description{
Title
}
