% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{tibble}
\name{mydataset}
\alias{mydataset}
\title{mydataset}
\description{
mydataset contains three regressors and one simulated response variable
with known parameter through a linear regression model with intercept
}
\author{
Francesco Barile \email{f.barile2@campus.unimb.it}
}
\keyword{data}
