% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets_doc.R
\docType{tibble}
\name{mydataset_noconst}
\alias{mydataset_noconst}
\title{mydataset_noconst}
\description{
mydataset_noconst contains three regressors and one simulated response variable
with known parameter through a linear regression model without intercept
}
\author{
Francesco Barile \email{f.barile2@campus.unimb.it}
}
\keyword{data}
