% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{y_pred}
\alias{y_pred}
\title{Function for predicting values}
\usage{
y_pred(formula, opt_obj, df)
}
\arguments{
\item{formula}{[character] a symbolic description of the model to be fitted (as lm), with -1 if the intercept is not included}

\item{opt_obj}{[list] output of either the function lm_GD_optimizer or lm_SD_optimizer}

\item{df}{[numeric] dataset containing the variables of interest}
}
\value{
[numeric] vector of the predicted values of the response variable
}
\description{
Function that uses the model parameters estimated and predict y starting from x
}
