% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betahat_GD_Cpp}
\alias{betahat_GD_Cpp}
\title{Estimation Linear Regression model via Gradient Descend method}
\arguments{
\item{beta}{initial guess beta vector}

\item{X}{design matrix}

\item{Y}{response variable vector}

\item{tolerance}{level}

\item{maxit}{max num of iter}

\item{stepsize}{learning parameter}
}
\description{
Estimation Linear Regression model via Gradient Descend method
}
