% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescent_R.R
\name{betahat_GD_R}
\alias{betahat_GD_R}
\title{betahat_GD_R}
\usage{
betahat_GD_R(beta, x, y, tolerance, maxit, stepsize)
}
\arguments{
\item{beta}{numeric vector containing an initial guess for the beta vector}

\item{x}{design matrix}

\item{y}{response variable vector}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}

\item{stepsize}{learning parameter}
}
\value{

}
\description{
This function estimates the beta vector of a linear Regression model via 
gradient descend method
}
