% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{betahat_SD_Cpp}
\alias{betahat_SD_Cpp}
\title{betahat_SD_Cpp}
\usage{
betahat_SD_Cpp(beta, X, Y, tolerance)
}
\arguments{
\item{beta}{initial guess beta vector}

\item{X}{design matrix}

\item{Y}{response variable vector}

\item{tolerance}{level}
}
\description{
Estimation Linear Regression model via Steepest Descend method
}
