% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SteepestDescent_R.R
\name{betahat_SD_R}
\alias{betahat_SD_R}
\title{Title}
\usage{
betahat_SD_R(beta, x, y, tolerance, maxit)
}
\arguments{
\item{beta}{numeric vector containing an initial guess for the beta vector}

\item{x}{design matrix}

\item{y}{response variable vector}

\item{tolerance}{tolerance level}

\item{maxit}{maximun number of iteration}
}
\value{

}
\description{
Title
}
