% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD_k fold cv_sequential.R
\name{cvsequencial}
\alias{cvsequencial}
\title{Sequential method to k-fold crossing validation}
\usage{
cvsequencial(k, formula, df, tolerance, maxit, stepsize, verbose)
}
\arguments{
\item{k}{[numeric] number of fold to create}

\item{formula}{[character] a symbolic description of the model to be fitted (as lm), with -1 if the intercept is not included}

\item{df}{[numeric] dataset containing the variables of interest}

\item{tolerance}{[numeric] tolerance level, stopping criteria of the algorithm (error<tolerance: stop)}

\item{maxit}{[numeric] maximum number of iterations, used if the stopping criteria is never matched}

\item{stepsize}{[numeric] learning parameter, to update the parameters at each iteration}

\item{verbose}{[logical] indicating if the function write messages during the computation}
}
\value{
[numeric] a vector of sample mean square errors calculated
}
\description{
cvsequencial is used to parallelize the k-fold cross validation
}
