% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_xy.R
\name{data_xy}
\alias{data_xy}
\title{Title}
\usage{
data_xy(formula, df)
}
\arguments{
\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}
}
\value{

}
\description{
Title
}
