% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_specification.R
\name{lm.model_spec}
\alias{lm.model_spec}
\title{Model Specification for LM}
\usage{
lm.model_spec(formula, df)
}
\arguments{
\item{formula}{a formula specifying the model}

\item{df}{dataset containing the variables of interest}
}
\value{
[list] of variable names and intercept (if included)
}
\description{
lm.model_spec is used to obtain the model specification as require for estimating a linear regression model
}
