% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_SD_optimizer.R
\name{lm_SD_optimizer}
\alias{lm_SD_optimizer}
\title{Estimating Linear Models via Steepest Descent Optimization method (General)}
\usage{
lm_SD_optimizer(formula, df, tolerance, maxit, verbose)
}
\arguments{
\item{formula}{[character] a symbolic description of the model to be fitted (as lm), with -1 if the intercept is not included}

\item{df}{[numeric] dataset containing the variables of interest}

\item{tolerance}{[numeric] tolerance level, stopping criteria of the algorithm (error<tolerance: stop)}

\item{maxit}{[numeric] maximum number of iterations, used if the stopping criteria is never matched}

\item{verbose}{[logical] indicating if the function write messages during the computation}
}
\value{
[list] betas estimated, regressors, response variable, formula of the model, and name of the method implemented
}
\description{
lm_SD_optimizer is used to estimates the parameters of a Linear Regression model via 
an unconstrained optimization method called Steepest Descent Method. It is in the general version: 
extract the variables of interest from the dataset and generates the initial values of the coefficients other than the betahat_SD_R functionalities
}
