% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\name{y_pred}
\alias{y_pred}
\title{Title}
\usage{
y_pred(formula, opt_obj, df)
}
\arguments{
\item{opt_obj}{output of either the function lm_GD_optimizer or lm_SD_optimizer}
}
\value{

}
\description{
Title
}
